requirements.txt
aenum==3.1.15
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Deprecated==1.2.14
Django==5.0
frozenlist==1.4.1
future==0.18.3
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
line-bot-sdk==3.5.1
multidict==6.0.4
numpy==1.26.2
oauthlib==3.2.2
pandas==2.1.4
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pygsheets==2.0.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.4
